Move plugin to "packages" folder inside x-pack
