+++ x-pack/packages/poc_kibana_cloud_plugin/tsconfig.json
