+++ x-pack/packages/poc_kibana_cloud_plugin/tslint.yaml
