+++ x-pack/plugins/poc_kibana_cloud_plugin_integration/public/app.js
@@ -7,13 +7,13 @@
 /* eslint-disable import/no-extraneous-dependencies */
 
 import React from 'react';
-// import { App as PocKibanaCloudPlugin } from '../../poc_kibana_cloud_plugin';
+import { App as PocKibanaCloudPlugin } from '../../../packages/poc_kibana_cloud_plugin';
 
 /**
  * After building the package, create a "poc-kibana-cloud-plugin" folder
  * and copy the "dist" folder generated, along with the package.json file in it
  * Then move this folder inside the node_modules folder.
  */
-import { App as PocKibanaCloudPlugin } from 'poc-kibana-cloud-plugin';
+// import { App as PocKibanaCloudPlugin } from 'poc-kibana-cloud-plugin';
 
 export const App = () => <PocKibanaCloudPlugin />;
